refactor + add delete dialog ui
